+++ CHANGELOG.rst
@@ -17,6 +17,9 @@ new version (on deck)
 - [ENH] Improvements to `pivot_longer` function, with improved speed and cleaner code. @samukweku
 - [ENH] Add kwargs to ``encode_categorical`` function, to create ordered categorical columns,
   or categorical columns with explicit categories. @samukweku
+- [ENH] Improvements to `complete` method. Use `pd.merge` to handle duplicates and
+  null values. @samukweku
+
 
 v0.20.10
 ========